backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2
supabase==2.5.0

# Task Queue
celery[redis]==5.4.0
redis==5.0.6

# AI/LLM
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-anthropic>=0.1.0,<0.2.0
openai>=1.0.0
anthropic>=0.30.0
tiktoken>=0.7.0

# Vector Store
chromadb>=0.5.0

# API Utilities
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# External APIs
requests>=2.31.0,<3.0.0
beautifulsoup4==4.12.3
google-search-results==2.4.2

# Email
resend>=1.0.0

# Monitoring
sentry-sdk==2.7.1
prometheus-client==0.20.0

# Development
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
black==24.4.2
ruff==0.5.0
mypy==1.10.1
pre-commit==3.7.1

# Type stubs
types-requests
types-redis